Update README.md
This is the repository of the STATS 401 Final Project of Group 3 including Xi Chen, Tianshu Feng, and Yuan Shan. Our topic is football transfer market visualization. The code folder stores the Python code to scrape our dataset and clean data for visualization. The scraped, raw data is stored in the data folder along with a combined version. The dashboard folder contains our visualization dashboard as well as essential processed data for our visualization.